requirements.txt
asgiref==3.2.7
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-writable-nested==0.6.0
drf-yasg==1.17.1
firebase-admin==4.2.0
gcloud==0.17.0
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.17.3
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
oauth2client==4.1.3
packaging==20.3
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
PyJWT==1.7.1
pyparsing==2.4.7
Pyrebase==3.0.27
python-jwt==3.2.6
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
